+++ src/main/resources/db/migration/postgresql/V2.7.7.20200430193000__add_amount_files_in_analysis_result.sql
@@ -1,1 +1,1 @@
-ALTER TABLE ${ohdsiSchema}.ee_analysis_status ADD amount_files_in_analysis BIGINT;
+ALTER TABLE ${ohdsiSchema}.ee_analysis_status ADD amount_files_in_analysis INTEGER;
